refactoring

+++ app/src/main/java/com/example/gnssnavigationstatus/model/Satellite.kt
@@ -1,3 +1,3 @@
 package com.example.gnssnavigationstatus.model
 
-data class Satellite(var id: Int, var type: String, var azim: Int, var elev: Int, var signalQuality: Int)
+data class Satellite(var id: Int, var type: String, var azim: Int, var elev: Int, var signalQuality: Float)
